+++ src/idunno.Bluesky/Unspecced/BlueskyAgent.cs
@@ -14,7 +14,7 @@ namespace idunno.Bluesky;
 public partial class BlueskyAgent
 {
     /// <summary>
-    /// Get an unspecced, paged list of views over globally popular feed generators.
+    /// Gets the age assurance state for the current user.
     /// </summary>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
@@ -46,6 +46,7 @@ public partial class BlueskyAgent
     /// <param name="query">Search query string; syntax, phrase, boolean, and faceting is unspecified, but Lucene query syntax is recommended.</param>
     /// <param name="limit">The maximum number of feed generators to return.</param>
     /// <param name="cursor">An optional cursor for pagination.</param>
+    /// <param name="subscribedLabelers">An optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
     /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt;100.</exception>
@@ -54,6 +55,7 @@ public partial class BlueskyAgent
         string? query = null,
         int? limit = null,
         string? cursor = null,
+        IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
     {
         if (limit is not null)
@@ -72,6 +74,7 @@ public partial class BlueskyAgent
             httpClient: HttpClient,
             onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
             loggerFactory: LoggerFactory,
+            subscribedLabelers: subscribedLabelers,
             maximumResponseSize: MaximumResponseSize,
             cancellationToken: cancellationToken).ConfigureAwait(false);
 #pragma warning restore BSKYUnspecced
@@ -80,13 +83,15 @@ public partial class BlueskyAgent
     /// <summary>
     /// Get a <see cref="ICollection{T}"/> of <see cref="StarterPackView"/>s of suggested starter packs.
     /// </summary>
-    /// <param name="limit">The number of starter packs to return. Must be between 1 and 50.</param>
+    /// <param name="limit">The number of starter packs to return. Must be between 1 and 25.</param>
+    /// <param name="subscribedLabelers">An optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt;50.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt;25.</exception>
     [Experimental("BSKYUnspecced", UrlFormat = "https://bluesky.idunno.dev/docs/unspecced.html")]
     public async Task<AtProtoHttpResult<ICollection<StarterPackView>>> GetSuggestedStarterPacks(
         int? limit = null,
+        IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
     {
         if (limit is not null)
@@ -103,6 +108,7 @@ public partial class BlueskyAgent
             httpClient: HttpClient,
             onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
             loggerFactory: LoggerFactory,
+            subscribedLabelers: subscribedLabelers,
             maximumResponseSize: MaximumResponseSize,
             cancellationToken: cancellationToken).ConfigureAwait(false);
 #pragma warning restore BSKYUnspecced
@@ -111,12 +117,14 @@ public partial class BlueskyAgent
     /// <summary>
     /// Get a collection of tagged suggestions.
     /// </summary>
-    /// <param name="parameters">Any parameters to send to the endpoint. Parameter values will automatically be query string encoded.</param>
+    /// <param name="parameters">Any parameters to send to the endpoint. Parameter names and values will automatically be query string encoded.</param>
+    /// <param name="subscribedLabelers">An optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
     [Experimental("BSKYUnspecced", UrlFormat = "https://bluesky.idunno.dev/docs/unspecced.html")]
     public async Task<AtProtoHttpResult<ICollection<Suggestion>>> GetTaggedSuggestions(
         ICollection<KeyValuePair<string, object>>? parameters = null,
+        IEnumerable<Did>? subscribedLabelers = null,
         CancellationToken cancellationToken = default)
     {
 #pragma warning disable BSKYUnspecced
@@ -127,6 +135,7 @@ public partial class BlueskyAgent
             httpClient: HttpClient,
             onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
             loggerFactory: LoggerFactory,
+            subscribedLabelers: subscribedLabelers,
             maximumResponseSize: MaximumResponseSize,
             cancellationToken: cancellationToken).ConfigureAwait(false);
 #pragma warning restore BSKYUnspecced
@@ -136,11 +145,12 @@ public partial class BlueskyAgent
     /// Get a collection of trending topics
     /// </summary>
     /// <param name="limit">The number of topics to return. Must be between 1 and 25.</param>
+    /// <param name="subscribedLabelers">An optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
     /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt;25.</exception>
     [Experimental("BSKYUnspecced", UrlFormat = "https://bluesky.idunno.dev/docs/unspecced.html")]
-    public async Task<AtProtoHttpResult<TrendingTopics>> GetTrendingTopics(int? limit = null, CancellationToken cancellationToken = default)
+    public async Task<AtProtoHttpResult<TrendingTopics>> GetTrendingTopics(int? limit = null, IEnumerable<Did>? subscribedLabelers = null, CancellationToken cancellationToken = default)
     {
         if (limit is not null)
         {
@@ -156,6 +166,7 @@ public partial class BlueskyAgent
             httpClient: HttpClient,
             onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
             loggerFactory: LoggerFactory,
+            subscribedLabelers: subscribedLabelers,
             maximumResponseSize: MaximumResponseSize,
             cancellationToken: cancellationToken).ConfigureAwait(false);
 #pragma warning restore BSKYUnspecced
@@ -165,11 +176,12 @@ public partial class BlueskyAgent
     /// Get a collection of trends.
     /// </summary>
     /// <param name="limit">The number of topics to return. Must be between 1 and 25.</param>
+    /// <param name="subscribedLabelers">An optional list of labeler <see cref="Did"/>s to accept labels from.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
     /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt;25.</exception>
     [Experimental("BSKYUnspecced", UrlFormat = "https://bluesky.idunno.dev/docs/unspecced.html")]
-    public async Task<AtProtoHttpResult<ICollection<TrendView>>> GetTrends(int? limit = null, CancellationToken cancellationToken = default)
+    public async Task<AtProtoHttpResult<RecommendationReadOnlyCollection<TrendView>>> GetTrends(int? limit = null, IEnumerable<Did>? subscribedLabelers = null, CancellationToken cancellationToken = default)
     {
         if (limit is not null)
         {
@@ -185,6 +197,7 @@ public partial class BlueskyAgent
             httpClient: HttpClient,
             onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
             loggerFactory: LoggerFactory,
+            subscribedLabelers: subscribedLabelers,
             maximumResponseSize: MaximumResponseSize,
             cancellationToken: cancellationToken).ConfigureAwait(false);
 #pragma warning restore BSKYUnspecced